refactor(address radio button card): rename file to use kebab case
