+++ stories/components/address-radio-button-card.stories.js
@@ -1,4 +1,4 @@
-import AddressRadioButtonCard from '../../components/AddressRadioButtonCard';
+import AddressRadioButtonCard from '../../components/address-radio-button-card';
 
 export default {
 	title: 'AddressRadioButtonCard',
